requirements.txt
discord.py>=2.3,<3.0
APScheduler>=3.10,<4.0
